refactor: extract monorepo dependency cascade into its own module
